message clean code
