+++ Airelax.Application/Houses/HouseAppService.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Net;
 using System.Threading.Tasks;
 using Airelax.Application.Account;
-using Airelax.Application.Helpers;
+//using Airelax.Application.Helpers;
 using Airelax.Application.Houses.Dtos.Request;
 using Airelax.Application.Houses.Dtos.Response;
 using Airelax.Domain.Comments;
